+++ src/main/kotlin/nn/learn/HebbLearner.kt
@@ -1,17 +1,16 @@
 package nn.learn
 
-import koma.matrix.ejml.backend.times
 import nn.DataVector
 import nn.Net
 
 class HebbLearner(private val net: Net,
-                  private var learningRate: Float)
+                  private var learningRate: Double)
 {
     fun learnSample(input: DataVector, label: DataVector)
     {
         val layer = this.net.layers[0]
         val output = layer.activation.forward(layer.forward(input))
-        val diff = (label - output) * this.learningRate.toDouble()
+        val diff = (label - output) * this.learningRate
         val deltas = diff * input
 
         layer.weights += deltas
